Initial Check In
This change implements a Naive Bayes classifier for English text documents. It includes a simple English tokenizer, and the classifier is powered with term frequency-inverse document frequency data for the word features.

splitDataSet.js and precalculate.js offer methods to set up training and test set data and to pretrain the classifer, respectively.