Implement web server
This implements a web server which supports accepting JSON data from outside origins.

The only method this handles, other than the CORS preflight OPTIONS, is POST on the root URL. The data sent must be a JSON array of objects in which each object has a string property, "full_text".

The return is the same set of objects but with a "category" property added which contains the string representing the category that the service classified the document as.